Use client_button instead of objection_button
Currently, we use objection_button class for the shouts, but the
toggleShout function actually changes it to client_button
instead. This makes it change style (height) whenever
it is pressed. We can fix this by just using client_button.
That makes it look more like the rest of the buttons as well.

Since we don't use objection_button anymore, remove it from
the styles.