Redo 2factor control flow
It appears that if 2factor is started, the client sends the HI
with the current fingerprint (as hdid), but in the 2fa block
it also sends the cookie hdid with no header. It seems unlikely
that the server can parse this packet correctly, so this looks
like a mistake. Remove this sendServer line and only do the
2factor flow when it's enabled, do not send handshake to the
server otherwise.